@@ -7,12 +7,19 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
-    "core-js": "^3.8.3",
-    "vue": "^2.6.14"
+    "@vue/cli": "^5.0.4",
+    "axios": "^0.20.0-0",
+    "better-scroll": "^2.4.2",
+    "core-js": "^3.6.5",
+    "element-ui": "^2.15.6",
+    "force": "^0.0.3",
+    "mitt": "^3.0.0",
+    "vue": "^3.2.25",
+    "vue-router": "^4.0.14"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "vue-template-compiler": "^2.6.14"
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-service": "^4.5.17",
+    "vue-template-compiler": "^2.6.11"
   }
 }
